updates to tables in testing

+++ src/sql/sales_raw.sql
@@ -1,5 +1,5 @@
 CREATE UNLOGGED TABLE sales_raw (
     id TEXT
     ,d TEXT
-    ,quantity INT
+    ,quantity SMALLINT
 )
